+++ internal/testsuite/class_c_test.go
@@ -252,6 +252,7 @@ func TestClassCScenarios(t *testing.T) {
 						Convey("Then the expected frame was sent", func() {
 							So(common.Gateway.(*test.GatewayBackend).TXPacketChan, ShouldHaveLength, 1)
 							txPacket := <-common.Gateway.(*test.GatewayBackend).TXPacketChan
+							So(txPacket.Token, ShouldNotEqual, 0)
 							So(&txPacket.TXInfo, ShouldResemble, t.ExpectedTXInfo)
 						})
 					} else {